T5 09/03 Pro duc xong

+++ 19T1021252/19T1021252.BusinessLayers/ProductDataService.cs
@@ -162,7 +162,7 @@ namespace _19T1021252.BusinessLayers
         /// </summary>
         /// <param name="attributeID"></param>
         /// <returns></returns>
-        public static ProductAttribute GetAttribute(int attributeID)
+        public static ProductAttribute GetAttribute(long attributeID)
         {
             return productDB.GetAttribute(attributeID);
         }
